+++ Gruntfile.js
@@ -17,7 +17,8 @@ module.exports = function(grunt) {
                 '<%= grunt.template.today("dd.mm.yyyy") %>\n' +
                 ' * <%= pkg.homepage %>\n' +
                 ' * Copyright (c) <%= grunt.template.today("yyyy") %> <%= pkg.author.name %>, <%= _.pluck(pkg.contributors, "name").join(", ") %>\n' +
-                ' * Licensed under the <%= pkg.license %> license\n */\n\n'
+                ' * Licensed under the <%= pkg.license %> license\n' +
+                ' * MODIFIED for i12 GmbH by Patrick Schaper December 2015\n */\n\n'
         },
 
         browserify: {
@@ -27,58 +28,14 @@ module.exports = function(grunt) {
                 },
                 banner: '<%= meta.banner %>'
             },
-            dev_demo: {
-                options: {
-                    browserifyOptions: {
-                        debug: true
-                    },
-                    keepAlive: true,
-                    watch: true
-                },
-                src: 'src/js/shariff.js',
-                dest: 'demo/app.min.js'
-            },
-            dist_complete_min: {
-                options: {
-                    transform: [ ['uglifyify', { global: true } ] ]
-                },
-                src: 'src/js/shariff.js',
-                dest: 'build/shariff.complete.js'
-            },
             dist_min: {
                 options: {
                     transform: [
                         ['uglifyify', { global: true } ],
-                        ['browserify-shim', { global: true } ]
                     ]
                 },
                 src: 'src/js/shariff.js',
                 dest: 'build/shariff.min.js'
-            },
-            demo: {
-                options: {
-                    transform: [ ['uglifyify', { global: true } ] ],
-                    watch: true
-                },
-                src: 'src/js/shariff.js',
-                dest: 'demo/app.min.js'
-            }
-        },
-
-        copy: {
-            demo: {
-                files: [
-                    {
-                        expand: true,
-                        cwd: 'node_modules/font-awesome',
-                        src: '{fonts,css}/*',
-                        dest: 'demo/'
-                    },
-                    {
-                        src: 'build/*',
-                        dest: 'demo/'
-                    }
-                ]
             }
         },
 
@@ -96,7 +53,6 @@ module.exports = function(grunt) {
             options: {
                 banner: '<%= meta.banner %>',
                 paths: [
-                    'node_modules/font-awesome/less',
                     'node_modules/shariff/src/style'
                 ],
                 plugins: [
@@ -105,76 +61,19 @@ module.exports = function(grunt) {
                 ],
                 strictMath: true
             },
-            demo: {
-                options: {
-                    sourceMap: true,
-                    outputSourceFiles: true,
-                    sourceMapFileInline: true,
-                    plugins: [
-                        new (require('less-plugin-autoprefix'))({
-                            browsers: browsers,
-                            map: true
-                        }),
-                        new (require('less-plugin-clean-css'))()
-                    ],
-                },
-                src: 'src/style/demo.less',
-                dest: 'demo/app.min.css'
-            },
-            dist: {
-                options: {
-                    modifyVars: {
-                        'fa-font-path': '"https://netdna.bootstrapcdn.com/font-awesome/4.3.0/fonts"'
-                    }
-                },
-                src: 'src/style/shariff-complete.less',
-                dest: 'build/shariff.complete.css'
-            },
             dist_min: {
                 src: 'src/style/shariff.less',
                 dest: 'build/shariff.min.css'
             }
         },
-
-        hapi: {
-            shariff: {
-                options: {
-                    server: require('path').resolve('./node_modules/shariff-backend-node/server.js'),
-                    // noasync: true,
-                }
-            }
-        },
-
-        connect: {
-            demo: {
-                options: {
-                    hostname: '0.0.0.0',
-                    // hostname: 'localhost',
-                    port: 8080,
-                    base: 'demo',
-                    keepalive: true,
-                    // livereload: true,
-                    // open: true,
-                    // debug: true,
-                    middleware: function (connect, options, middlewares) {
-                        var proxy = require('grunt-connect-proxy/lib/utils').proxyRequest;
-                        return [
-                            proxy,
-                            connect.static(options.base[0]),
-                            connect.directory(options.base[0])
-                        ];
-                    }
-                },
-                proxies: [
+        file_append: {
+            default_options: {
+                files: [
                     {
-                        rewrite: {
-                            '^/shariff': ''
-                        },
-                        context: '/shariff/',
-                        host: 'localhost',
-                        port: 3001,
-                        https: false,
-                        xforward: false
+                        prepend: '(function($) { \n\t$(document).ready(function() {\n',
+                        append: '\n\t});\n})(cms_jQuery);',
+                        input: 'build/shariff.min.js',
+                        output: 'build/shariff.min.js'
                     }
                 ]
             }
@@ -182,15 +81,10 @@ module.exports = function(grunt) {
     });
 
     grunt.loadNpmTasks('grunt-browserify');
-    grunt.loadNpmTasks('grunt-contrib-copy');
     grunt.loadNpmTasks('grunt-contrib-jshint');
     grunt.loadNpmTasks('grunt-contrib-less');
-    grunt.loadNpmTasks('grunt-contrib-connect');
-    grunt.loadNpmTasks('grunt-connect-proxy');
-    grunt.loadNpmTasks('grunt-hapi');
+    grunt.loadNpmTasks('grunt-file-append');
 
     grunt.registerTask('test', ['jshint']);
-    grunt.registerTask('build', ['test', 'less:demo', 'less:dist', 'less:dist_min', 'browserify:dist_complete_min', 'browserify:dist_min']);
-    grunt.registerTask('demo', ['copy:demo', 'less:demo', 'browserify:demo', 'hapi', 'configureProxies:demo', 'connect']);
-    grunt.registerTask('default', ['test', 'demo']);
+    grunt.registerTask('build', ['test', 'less:dist_min', 'browserify:dist_min', 'file_append']);
 };