+++ release.sh
@@ -14,9 +14,8 @@ git push
 7z a -tzip $ZIP ./build/*
 # rm -fr build
 # https://github.com/aktau/github-release
-github-release release -u heiseonline -r shariff -t $VVERSION -n $VERSION
-github-release upload -u heiseonline -r shariff -t $VVERSION -n $ZIP -f $ZIP
+github-release release -u yesman82 -r shariff -t $VVERSION -n $VERSION
+github-release upload -u yesman82 -r shariff -t $VVERSION -n $ZIP -f $ZIP
 git push
 git push --tags
 rm $ZIP
-npm publish